Restructured project structure and introduced model class for ip
