+++ src/main/kotlin/com/kiwi/to/the/moon/controller/HueController.kt
@@ -1,5 +1,6 @@
-package com.kiwi.to.the.moon
+package com.kiwi.to.the.moon.controller
 
+import com.kiwi.to.the.moon.service.HueService
 import javax.inject.Inject
 import javax.ws.rs.GET
 import javax.ws.rs.Path
@@ -11,7 +12,7 @@ import javax.ws.rs.Consumes
 @Path("/hue")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-class HueResource {
+class HueController {
 
     @Inject
     lateinit var hueService: HueService
