+++ src/main/kotlin/com/kiwi/to/the/moon/model/HueAttributes.kt
@@ -1,0 +1,3 @@
+package com.kiwi.to.the.moon.model
+
+data class HueAttributes(val ip: String?)
