+++ src/main/resources/application.properties
@@ -1,2 +1,2 @@
 hue.name=CatHue
-hue.apiKey=
+hue.apiKey=4QJL6MSzsyk6ZP0Ic-hIcHjQpfxYdBHZIT6ny9Zt
