+++ src/native-test/kotlin/com/kiwi/to/the/moon/NativeHueControllerIT.kt
@@ -3,4 +3,4 @@ package com.kiwi.to.the.moon
 import io.quarkus.test.junit.NativeImageTest
 
 @NativeImageTest
-class NativeHueResourceIT : HueResourceTest()
+class NativeHueControllerIT : HueControllerTest()
